moving around, not completely sure where going with this :)
