+++ local-db.bat
@@ -1,0 +1,2 @@
+
+docker run --rm --name marketobserver-mysql -p 3306:3306 -e MYSQL_DATABASE=marketobserver -e MYSQL_ROOT_PASSWORD=squirrels mysql:8
