+++ src/main/java/net/redirectme/squirrelacademy/marketobserver/character/data/PublicInfoRepository.java
@@ -1,4 +1,4 @@
-package net.redirectme.squirrelacademy.character.data;
+package net.redirectme.squirrelacademy.marketobserver.character.data;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
