+++ src/main/java/net/redirectme/squirrelacademy/marketobserver/market/MarketApi_Functions.java
@@ -1,15 +1,13 @@
-package net.redirectme.squirrelacademy.market;
+package net.redirectme.squirrelacademy.marketobserver.market;
 
 import org.openapitools.client.model.GetMarketsStructuresStructureId200Ok;
 import org.springframework.beans.factory.annotation.Value;
-import org.springframework.context.annotation.PropertySource;
 import org.springframework.stereotype.Service;
 
 import java.util.ArrayList;
 import java.util.List;
 
 @Service
-@PropertySource("classpath:eveesi.properties")
 public class MarketApi_Functions {
     @Value("${esidatasource}")
     String esidatasource; // ="tranquility";
