+++ src/main/java/net/redirectme/squirrelacademy/marketobserver/oauth2/MyOAuth2AuthorizedClientservice.java
@@ -1,4 +1,4 @@
-package net.redirectme.squirrelacademy.MarketObserver.oauth2;
+package net.redirectme.squirrelacademy.marketobserver.oauth2;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.jdbc.core.JdbcOperations;
