+++ src/main/resources/application.properties
@@ -1,0 +1,50 @@
+spring.session.store-type=jdbc
+spring.session.jdbc.initialize-schema=always
+spring.session.timeout.seconds=-1
+spring.session.jdbc.table-name=SPRING_SESSION
+
+# DATABASE================
+spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
+# spring.datasource.url=jdbc:mysql://squirrelacademy.redirectme.net:3306/marketobserver?createDatabaseIfNotExist=true&useSSL=false&rewriteBatchedStatements=true
+spring.datasource.url=jdbc:mysql://localhost:3306/marketobserver
+spring.datasource.username=root
+spring.datasource.password=squirrels
+
+# JPA/Hibernate================
+#or create
+spring.jpa.hibernate.ddl-auto=update
+spring.jpa.show-sql=false
+spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL5Dialect
+
+# EVE SSO
+#security.oauth2.client.id=eve
+#security.oauth2.client.grant-type=authorization_code=eve
+#security.oauth2.client.user-authorization-uri=https://login.eveonline.com/v2/oauth/authorize
+#security.oauth2.client.access-token-uri=https://login.eveonline.com/v2/oauth/token
+#security.oauth2.client.pre-established-redirect-uri=http://localhost:5000/login/oauth2/code/
+#security.oauth2.client.use-current-uri=false
+#security.oauth2.client.scope=publicData,esi-markets.structure_markets.v1,esi-characters.read_standings.v1
+#security.oauth2.client.authenticationScheme=query
+#security.oauth2.client.clientAuthenticationScheme=form
+#security.oauth2.resource.user-info-uri=https://login.eveonline.com/oauth/verify
+#security.oauth2.resource.token-info-uri=https://login.eveonline.com/oauth/verify
+
+spring.security.oauth2.client.registration.custom.client-id=5e4f55829ef74a839448c6532b218bd2
+spring.security.oauth2.client.registration.custom.client-secret=iqeBEZ1HLO0T5NGXxO58KMfqy5h3YV8t7wZWbFCp
+spring.security.oauth2.client.registration.custom.authorization-grant-type=authorization_code
+spring.security.oauth2.client.registration.custom.redirect-uri=http://localhost:8080/login/oauth2/code/custom
+spring.security.oauth2.client.registration.custom.scope=publicData,esi-markets.structure_markets.v1,esi-characters.read_standings.v1
+spring.security.oauth2.client.registration.custom.provider=custom
+
+spring.security.oauth2.client.provider.custom.authorization-uri=https://login.eveonline.com/v2/oauth/authorize/
+spring.security.oauth2.client.provider.custom.token-uri=https://login.eveonline.com/v2/oauth/token
+
+esidatasource = tranquility
+locationid_default=1027847403762
+locationid_ks=1028148435908
+language=en-us
+allowed_corporations=1441288013,1000111
+
+api_character_standings=https://esi.evetech.net/latest/characters/{character_id}/standings
+api_character_publicInfo=https://esi.evetech.net/latest/characters/{character_id}
+api_market_structures=https://esi.evetech.net/latest/markets/structures/{location_id}
