--- src/main/resources/application_nokeys.properties
@@ -1,18 +1,0 @@
-spring.session.store-type=jdbc
-spring.session.jdbc.initialize-schema=always
-spring.session.timeout.seconds=-1
-spring.session.jdbc.table-name=SPRING_SESSION
-
-
-
-# DATABASE================
-spring.datasource.driver-class-name=com.mysql.cj.jdbc.Driver
-spring.datasource.url=jdbc:mysql://squirrelacademy.redirectme.net:3306/marketobserver?createDatabaseIfNotExist=true&useSSL=false&rewriteBatchedStatements=true
-spring.datasource.username=
-spring.datasource.password=
-
-# JPA/Hibernate================
-#or create
-spring.jpa.hibernate.ddl-auto=update
-spring.jpa.show-sql=false
-spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.MySQL5Dialect
