+++ src/main/resources/eveesi_nokeys.properties
@@ -4,7 +4,6 @@ locationid_ks=1028148435908
 language=en-us
 allowed_corporations=1441288013,1000111
 
-
 api_character_standings=https://esi.evetech.net/latest/characters/{character_id}/standings
 api_character_publicInfo=https://esi.evetech.net/latest/characters/{character_id}
 api_market_structures=https://esi.evetech.net/latest/markets/structures/{location_id}
